backend/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.11.1
sqlparse==0.5.3
